% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareBranchLength.R
\name{compareBranchLength}
\alias{compareBranchLength}
\title{compareBranchLength}
\usage{
compareBranchLength(
  tree1,
  tree2,
  write = NULL,
  outgroup = NULL,
  root = NULL,
  composition = F,
  unique = F
)
}
\arguments{
\item{tree1}{A phylo object without branch lengths}

\item{tree2}{A phylo object with branch lengths}

\item{write}{Optional. Specify the name of the dataframe file to be written locally (nothing is written if this parameter is not specified)}

\item{outgroup}{Optional. Specify outgroup taxa to remove (by default, outgroup = F assumes that the user does not want to remove outgroup taxa)}

\item{root}{Optional. Specify the same root for both trees, which is recommended to facilitate tree comparisons (by default, root = F assumes that trees share the same root)}

\item{composition.}{Optional. Specify if composition of corresponding clades should be present in the dataframe (by default, composition = F)}

\item{unique.}{Optional. Show unique clades (default: unique = F).}
}
\description{
Given two trees with branch lengths, return a dataframe containing lengths from matching branches.
}
\examples{
# Example 1 (identify unique nodes)
tree1 = read.tree (text="(t1,(t3,(t2,(t4,t5))));")
tree2 = read.tree (text="(t1,(t2,(t3,(t4,t5)47)53)94);")
compareBranchLength (tree1, tree2)

}
\author{
Daniel YM Nakamura
}
