% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterMissing.R
\name{filterMissing}
\alias{filterMissing}
\title{filterMissing}
\usage{
filterMissing(
  input,
  input_format = NULL,
  output,
  output_format = "nexus",
  missing = "row"
)
}
\arguments{
\item{input}{Input file (morphological matrix in Nexus format).}

\item{input_format}{To load from a local file: 'nexus' or 'tnt'.}

\item{output}{Output path (e.g. if the user specify it as "Desktop/Index", the output files will be "Desktop/Index_ORDERED.nexus" and "Desktop/Index_UNORDERED.nexus")}

\item{output_format}{Format to write the output matrix. Options: "nexus" (default) or "tnt".}

\item{missing}{Parameter specifying if rows and/or columns in which all cells are missing data (?) should be removed. Options: "row" (default i.e. terminals), "column" (i.e. characters, transformation series), "both".}
}
\description{
Filter taxa (rows) and/or characters (columns) in a matrix containing only missing data.
}
\examples{
\dontrun{
# Example
filterMissing(input = "testdata/test_filterMissing.nexus", output = "testdata/test", missing = "row")
}

}
\author{
Daniel YM Nakamura
}
