% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapBranchLength.R
\name{mapBranchLength}
\alias{mapBranchLength}
\title{mapBranchLength}
\usage{
mapBranchLength(
  tree1,
  tree2,
  write = NULL,
  outgroup = NULL,
  root = NULL,
  plotTrees = F,
  node.numbers = T,
  tree.width = 10,
  tree.height = 10,
  tree.fsize = 0.5,
  tree.adj = c(-1.5, 0.5),
  tree.cex = 2,
  tree.output = "trees_unique_nodes.pdf"
)
}
\arguments{
\item{tree1}{A phylo object without branch lengths}

\item{tree2}{A phylo object with branch lengths}

\item{write}{Optional. Specify the name of tree file to be written locally (nothing is written if this parameter is not specified)}

\item{outgroup}{Optional. Specify outgroup taxa to remove (by default, outgroup = F assumes that the user does not want to remove outgroup taxa)}

\item{root}{Optional. Specify the same root for both trees, which is recommended to facilitate tree comparisons (by default, root = F assumes that trees share the same root)}
}
\description{
Given a tree 1 without branch lengths (e.g. strict consensus of optimal trees) and a tree 2 with branch lengths (e.g. each MPT before reconciliation into a strict consensus), return tree 1 with branch lengths from shared clades with tree 2.
}
\examples{
# Example 1 (identify unique nodes)
tree1 = read.tree (text="(t1,(t3,(t2,(t4,t5))));")
tree2 = read.tree (text="(t1,(t2,(t3,(t4,t5)47)53)94);")
mapBranchLength (tree1, tree2)

}
\author{
Daniel YM Nakamura
}
