% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiCID.R
\name{multiCID}
\alias{multiCID}
\title{multiCID}
\usage{
multiCID(
  trees1,
  trees2,
  method = "random",
  normalization = FALSE,
  subsample = 1
)
}
\arguments{
\item{trees1}{A \code{phylo} or \code{multiPhylo} object with multiple trees that can be loaded using \code{ape::read.tree} for NEWICK files or \code{TreeTools::ReadTntTree} for TNT files. If the pool of MPTs presents binary and non-binary trees, only binary trees are processed.}

\item{trees2}{Another \code{phylo} or \code{multiPhylo} object.}

\item{method}{Optional. Specify if CID distances will be calculated by (1) \code{random} (default: selects one binary tree randomly from the multiPhylo object), (2) \code{meanCID} (calculates mean of all pairwise CID distances between two \code{multiPhylo} objects), (3) \code{minCID} (calculates the minimum), (4) \code{maxCID} (calculates the maximum), or (5) \code{all} (calculates mean, minimum, and maximum values).}

\item{normalization}{Optional. Specify if CID distances should be normalized. By default, CID distances are not normalized.}

\item{subsample}{Optional. Specify if CID distances will be calculated to only a fraction of the total number of trees available in each set (default = 1; i.e. all trees are evaluated). Zero values are not accepted.}
}
\description{
\code{multiCID} computes CID distances between two sets of binary trees (e.g. MPTs), \eqn{T_1 = \{\text{Tree}_1, \text{Tree}_2, \dots, \text{Tree}_n\}} and \eqn{T_2 = \{\text{Tree}_a, \text{Tree}_b, \dots, \text{Tree}_z\}}. The methods available are (1) randomly selecting one of the binary trees from each set (quick and naive) and (2) estimating the mean CID (or minimum or maximum) from \eqn{n} pairwise combinations between the two sets. Both trees must contain the same set of leaves.
}
\author{
Daniel YM Nakamura
}
