% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiRF.R
\name{multiRF}
\alias{multiRF}
\title{multiRF}
\usage{
multiRF(
  trees1,
  trees2,
  method = "random",
  normalization = FALSE,
  subsample = 1
)
}
\arguments{
\item{trees1}{A \code{phylo} or \code{multiPhylo} object with multiple trees that can be loaded using \code{ape::read.tree} for NEWICK files or \code{TreeTools::ReadTntTree} for TNT files. If the pool of MPTs presents binary and non-binary trees, only binary trees are processed.}

\item{trees2}{Another \code{phylo} or \code{multiPhylo} object.}

\item{method}{Optional. Specify if RF distances will be calculated by (1) \code{random} (default: selects one binary tree randomly from the multiPhylo object), (2) \code{meanRF} (calculates mean of all pairwise RF distances between two \code{multiPhylo} objects), (3) \code{minRF} (calculates the minimum), (4) \code{maxRF} (calculates the maximum), or (5) \code{all} (calculates mean, minimum, and maximum values).}

\item{normalization}{Optional. Specify if RF distances should be normalized using the maximum possible RF distance \eqn{i(T_1) + i(T_2)} (sum of clades in both trees). By default, RF distances are not normalized.}

\item{subsample}{Optional. Specify if RF distances will be calculated to only a fraction of the total number of trees available in each set (default = 1; i.e. all trees are evaluated). Zero values are not accepted.}
}
\description{
\code{multiRF} computes Robinson-Foulds distances between two sets of binary trees (e.g. MPTs), \eqn{T_1 = \{\text{Tree}_1, \text{Tree}_2, \dots, \text{Tree}_n\}} and \eqn{T_2 = \{\text{Tree}_a, \text{Tree}_b, \dots, \text{Tree}_z\}}. The methods available are (1) randomly selecting one of the binary trees from each set (quick and naive) and (2) estimating the mean RF (or minimum or maximum) from \eqn{n} pairwise combinations between the two sets. Both trees must contain the same set of leaves.
}
\author{
Daniel YM Nakamura
}
