% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiSPR.R
\name{multiSPR}
\alias{multiSPR}
\title{multiSPR}
\usage{
multiSPR(
  trees1,
  trees2,
  method = "random",
  normalization = FALSE,
  subsample = 1
)
}
\arguments{
\item{trees1}{A \code{phylo} or \code{multiPhylo} object with multiple trees that can be loaded using \code{ape::read.tree} for NEWICK files or \code{TreeTools::ReadTntTree} for TNT files. If the pool of MPTs presents binary and non-binary trees, only binary trees are processed.}

\item{trees2}{Another \code{phylo} or \code{multiPhylo} object.}

\item{method}{Optional. Specify if SPR distances will be calculated by (1) \code{random} (default: selects one binary tree randomly from the multiPhylo object), (2) \code{meanSPR} (calculates mean of all pairwise SPR distances between two \code{multiPhylo} objects) or (3) \code{minSPR} (calculates the minimum values of all pairwise SPR distances between two \code{multiPhylo} objects. The option \code{meanSPR} can be slow if the number of MPTs is high.}

\item{normalization}{Optional. Specify if SPR distances should be normalized using upper bound values (Ding et al. 2011). See details in \link{normalizedSPR}. By default, SPR distances are not normalized.}

\item{subsample}{Optional. Specify if SPR distances will be calculated to only a fraction of the total number of trees available in each set (default = 1; i.e. all trees are evaluated). Zero values are not accepted.}
}
\description{
\code{multiSPR} computes SPR distances between two sets of binary trees (e.g. MPTs), \eqn{T_1 = \{\text{Tree}_1, \text{Tree}_2, \dots, \text{Tree}_n\}} and \eqn{T_2 = \{\text{Tree}_a, \text{Tree}_b, \dots, \text{Tree}_z\}}. The methods available are (1) randomly selecting one of the binary trees from each set (quick and naive), (2) estimating the mean SPR from \eqn{n} pairwise combinations between the two sets, and (3) estimating the minimum SPR from \eqn{n} pairwise combinations between the two sets. This function is useful when the two strict consensus trees exhibit polytomies. Both trees must contain the same set of leaves.
}
\references{
Ding, Y., Grünewald, S., Humphries, P.J., 2011. On agreement forests. J. Comb. Theory Ser. 118(7), 2059–2065.
}
\author{
Daniel YM Nakamura
}
