% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nexus2tnt.R
\name{nexus2tnt}
\alias{nexus2tnt}
\title{nexus2tnt}
\usage{
nexus2tnt(input_file, output_file)
}
\arguments{
\item{input_file}{Path to the input NEXUS file.}

\item{output_file}{Path to the output TNT file.}
}
\description{
Converts a NEXUS matrix into TNT format while preserving header data type and ordered character structures.
}
\author{
Daniel YM Nakamura
}
