% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrodictNodes.R
\name{retrodictNodes}
\alias{retrodictNodes}
\title{retrodictNodes}
\usage{
retrodictNodes(tree1, tree2, outgroup = NULL, root = NULL, dataframe = F)
}
\arguments{
\item{tree1}{A .phylo tree that can be loaded using ape::read.tree for NEWICK files or TreeTools::ReadTntTree for TNT files. The .phylo must contain $node.label.}

\item{tree2}{Another .phylo tree.}

\item{outgroup}{Optional. Specify outgroup taxa to remove (by default, the function assumes that the user does not want to remove outgroup taxa)}

\item{root}{Optional. Specify the same root for both trees, which is recommended to facilitate tree comparisons (by default, the function assumes that trees share the same root)}

\item{dataframe}{Optional. Write a TSV file in current directory containing the output dataframe (by default, no .TSV is written).}
}
\description{
Given two trees with support values, creates a dataframe containing support values of tree 1 and the occurrence of the clade in tree 2. This dataframe can be used for logistic regressions testing whether support values of tree 1 retrodict the occurrence of clades in tree 2.
}
\examples{
# Example 1 (simplest case)
tree1 = read.tree (text="(t1,(t2,(t3,(t4,t5)75)32)45);")
tree2 = read.tree (text="(t1,(t6,(t3,(t4,t5)47)53)94);")
retrodictNodes (tree1, tree2)

}
\author{
Daniel YM Nakamura
}
