% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitOrdFromUnord.R
\name{splitOrdFromUnord}
\alias{splitOrdFromUnord}
\title{splitOrdFromUnord}
\usage{
splitOrdFromUnord(input, input_format, output_index, list_ordered)
}
\arguments{
\item{input}{Input file (concatenated morphological matrix loaded locally or already loaded in R).}

\item{input_format}{To load from a local file: 'nexus' or 'tnt'.}

\item{output_index}{Output index (e.g. if the user specify it as "Desktop/Index", the output files will be "Desktop/Index_ORDERED.nexus" and "Desktop/Index_UNORDERED.nexus")}

\item{list_ordered}{List of ordered characters e.g. c(1, 3, 9, 13). Character numbering starts with 1 (even if input data is .TNT).}
}
\description{
Splits a morphological matrix containing ordered and unordered characters into two matrices (ordered and unordered). This is useful to run phylogenetic analyses with MK and ORDERED models in IQTREE, especially if a concatenated matrix containing invariant characters is given as input.
}
\examples{
# Example
splitOrdFromUnord(input="../testdata/048_MORPH_data.nex", output_index = "../testdata/048_MORPH", list_ordered=list_ordered)

}
\author{
Daniel YM Nakamura
}
