% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tnt2nexus.R
\name{tnt2nexus}
\alias{tnt2nexus}
\title{tnt2nexus}
\usage{
tnt2nexus(input_file, output_file)
}
\arguments{
\item{input_file}{Path to the input TNT file.}

\item{output_file}{Path to the output NEXUS file.}
}
\description{
Converts a TNT matrix into NEXUS format while preserving header data type and ordered character structures.
}
\author{
Daniel YM Nakamura
}
