% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterInvariants.R
\name{filterInvariants}
\alias{filterInvariants}
\title{filterInvariants}
\usage{
filterInvariants(input, output_index)
}
\arguments{
\item{input}{Input file (molecular or morphological matrix in Nexus or already loaded as a matrix object in R).}

\item{output_index}{Output index (e.g. if the user specify it as "Desktop/Index", the output files will be "Desktop/Index_onlyVARIANTS.nexus")}
}
\description{
Delete invariant characters in a morphological matrix. We follow the definition of invariant site from IQ-Tree: (1) constant sites containing only a single character state in all sequences, (2) partially constant sites (N and/or -), and (3) ambiguously constant sites (e.g. C, Y and -).
}
\examples{
# Example
filterInvariants(input="../testdata/015_MORPH_data.nexus", output_index="../testdata/015_MORPH_data")

}
\author{
Daniel YM Nakamura
}
