% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapSupport.R
\name{mapSupport}
\alias{mapSupport}
\title{mapSupport}
\usage{
mapSupport(tree1, tree2, write = NULL, outgroup = NULL, root = NULL)
}
\arguments{
\item{tree1}{A phylo object without support values}

\item{tree2}{A phylo object with support values}

\item{write}{Optional. Specify the name of tree file to be written locally (nothing is written if this parameter is not specified)}

\item{outgroup}{Optional. Specify outgroup taxa to remove (by default, outgroup = F assumes that the user does not want to remove outgroup taxa)}

\item{root}{Optional. Specify the same root for both trees, which is recommended to facilitate tree comparisons (by default, root = F assumes that trees share the same root)}
}
\description{
Given a tree 1 without support values (e.g. strict consensus of optimal trees) and tree 2 with support values (e.g. majority consensus from suboptimal bootstrap pseudo-replicates), return tree 1 with support values from shared clades with tree 2.
}
\examples{
# Example 1 (identify unique nodes)
tree1 = read.tree (text="(t1,(t3,(t2,(t4,t5))));")
tree2 = read.tree (text="(t1,(t2,(t3,(t4,t5)47)53)94);")
mapSupport (tree1, tree2)

}
\author{
Daniel YM Nakamura
}
