% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharedNodes.R
\name{sharedNodes}
\alias{sharedNodes}
\title{sharedNodes}
\usage{
sharedNodes(
  tree1,
  tree2,
  composition = F,
  outgroup = NULL,
  root = NULL,
  plotTrees = F,
  width = NULL,
  height = NULL,
  fsize = NULL,
  adj = NULL,
  cex = NULL,
  dataframe = T,
  messages = T,
  spearman = T
)
}
\arguments{
\item{tree1}{A .phylo tree that can be loaded using ape::read.tree for NEWICK files or TreeTools::ReadTntTree for TNT files}

\item{tree2}{Another .phylo tree}

\item{composition}{Optional. Specify if composition of corresponding clades should be present in the dataframe (by default, composition = F)}

\item{outgroup}{Optional. Specify outgroup taxa to remove (by default, the function assumes that the user does not want to remove outgroup taxa)}

\item{root}{Optional. Specify the same root for both trees, which is recommended to facilitate tree comparisons (by default, the function assumes that trees share the same root)}

\item{plotTrees}{Optional. Plot the two trees after taxa pruning in PDF format. If plot=T, the user should also adjust PDF dimensions (e.g. width = 8, height = 8), label size (e.g. fsize = 4), and position and size of support values (e.g. adj = c(-1.5,0.5), cex = 0.6).}

\item{dataframe}{Optional. Write a TSV file in current directory containing the output dataframe (by default, no .TSV is written).}

\item{spearman}{Optional. Test the correlation between support values using a Spearman test (by default, spearman = T.}
}
\description{
Compare shared clades between two trees. The outputs are (1) basic statistics about number of shared clades and support values; (2) a dataframe with node labels, descendants, and support values of shared clades, which facilitates descriptive and statistical comparisons of clade composition and support between corresponding nodes.
}
\examples{
# Example 1 (simplest case)
tree1 = read.tree (text="(t1,(t2,(t3,(t4,t5)75)32)45);")
tree2 = read.tree (text="(t1,(t6,(t3,(t4,t5)47)53)94);")
sharedNodes (tree1, tree2)

# Example 2 (show internal topology of each node, remove outgroup taxa t9 and t8, and reroot in t1 in both trees)
sharedNodes (tree1, tree2, composition=T, outgroup=c("t9", "t8"), root="t1",)

# Example 3 (plot  two trees)
sharedNodes (tree1, tree2, plotTrees=T, width=8, height=8, fsize=3, adj=c(-1.5,0.5), cex=0.6)

}
\author{
Daniel YM Nakamura, Taran Grant
}
