% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharedNodes.R
\name{sharedNodes}
\alias{sharedNodes}
\title{sharedNodes}
\usage{
sharedNodes(
  tree1,
  tree2,
  composition = F,
  outgroup = NULL,
  root = NULL,
  plotTrees = F,
  node.numbers = T,
  tree.width = NULL,
  tree.height = NULL,
  tree.fsize = NULL,
  tree.adj = NULL,
  tree.cex = NULL,
  tree2.direction = "rightwards",
  output.tree1 = "tree1_pruned.pdf",
  output.tree2 = "tree2_pruned.pdf",
  tanglegram = F,
  output.tangletree = "tanglegram_comparison.pdf",
  tanglegram.margin = 7,
  tanglegram.lab.cex = 0.5,
  tanglegram.edge.lwd = c(0.1, 0.1),
  tanglegram.lwd = 1,
  tanglegram.axes = F,
  tanglegram.width = 5,
  tanglegram.height = 5,
  tanglegram.colors = T,
  dataframe = F,
  output.dataframe = "shared.nodes.tsv",
  messages = T,
  spearman = F,
  write.pruned = F,
  write.pruned1.name = "pruned_tree1.nwk",
  write.pruned2.name = "pruned_tree2.nwk"
)
}
\arguments{
\item{tree1}{A \code{phylo} tree that can be loaded using \code{ape::read.tree} for \code{NEWICK} files or \code{TreeTools::ReadTntTree} for \code{TNT} files. The \code{phylo} must contain the element \code{$node.label}.}

\item{tree2}{Another \code{phylo} tree}

\item{composition}{Optional. Specify if composition of corresponding clades should be present in the dataframe (by default, \code{composition = F})}

\item{outgroup}{Optional. Specify outgroup taxa to remove (by default, the function assumes that the user does not want to remove outgroup taxa; i.e. \code{outgroup = NULL})}

\item{root}{Optional. Specify the same root for both trees, which is recommended to facilitate tree comparisons (by default, the function assumes that trees share the same root; i.e. \code{root = NULL})}

\item{plotTrees}{Optional. Plot the two trees after taxa pruning in \code{PDF} format. If \code{plot = T}, the user should also adjust \code{PDF} dimensions (e.g. \code{width = 8}, \code{height = 8}), label size (e.g. \code{fsize = 4}), and position and size of support values (e.g. \code{adj = c(-1.5,0.5)}, \code{cex = 0.6}).}

\item{node.numbers}{Optional. If plotTrees = T, show node index (do not confuse with support values'by default, True).}

\item{tree.width}{Optional. Width of trees in PDF if plotTrees = T.}

\item{tree.height}{Optional. Height of trees in PDF if plotTrees = T.}

\item{tree.fsize}{Optional. Font size in PDF if plotTrees = T.}

\item{tree.adj}{Optional. Adjust horizontal and vertical position if plotTrees = T.}

\item{tree.cex}{Optional. Adjust support size in nodes if plotTrees = T.}

\item{tree2.direction}{Optional. Adjust the direction of tree2 ("leftwards" vs "rightwards") if plotTrees = T.}

\item{output.tree1}{Optional. The output file name of tree 1 if plotTrees = T.}

\item{output.tree2}{Optional. The output file name of tree 2 if plotTrees = T.}

\item{tanglegram}{Optional. Plot a tanglegram minimizing the number of crosses of lines linking two trees in \code{PDF} format. If the input tree has no branch length, uniform lengths are simulated to enable visualization.}

\item{output.tangletree}{Optional. The output file name if tanglegram = T.}

\item{tanglegram.margin}{Optional. Distance between tangle trees (default = 7) if tanglegram = T.}

\item{tanglegram.lab.cex}{Optional. Size of leaf names (default = 0.5) if tanglegram = T.}

\item{tanglegram.edge.lwd}{Optional. Thickness of edges of tangle trees (default = c(0.1, 0.1)) if tanglegram = T.}

\item{tanglegram.lwd}{Optional. Thickness of edges connecting both tangle trees (default = 1) if tanglegram = T.}

\item{tanglegram.axes}{Optional. Show scale in tangle trees (default = F) if tanglegram = T.}

\item{tanglegram.width}{Optional. Width of tangle trees in PDF (default = 5) if tanglegram = T.}

\item{tanglegram.height}{Optional. Height of tangle trees in PDF (default = 5) if tanglegram = T.}

\item{tanglegram.colors}{Optional. Show color in edges connecting both tangle trees (default = T) if tanglegram = T.}

\item{dataframe}{Optional. Write a \code{TSV} file in current directory containing the output dataframe (by default, no \code{TSV} is written).}

\item{output.dataframe}{Optional. The output file name of the dataframe if dataframe = T.}

\item{spearman}{Optional. Test the correlation between support values using a Spearman test (by default, \code{spearman = T}).}

\item{write.pruned}{Write the pruned trees.}

\item{write.pruned1.name}{Output file name of pruned tree1 if write.pruned = T.}

\item{write.pruned2.name}{Output file name of pruned tree2 if write.pruned = T.}
}
\description{
Compare support values of shared clades between two trees. The outputs are (1) basic statistics about number of shared clades and support values; (2) a dataframe with node labels, descendants, and support values of shared clades, which facilitates descriptive and statistical comparisons of clade composition and support between corresponding nodes.
}
\examples{
# Example 1 (simplest case)
tree1 = read.tree (text="(t1,(t2,(t3,(t4,t5)75)32)45);")
tree2 = read.tree (text="(t1,(t6,(t3,(t4,t5)47)53)94);")
sharedNodes (tree1, tree2)

# Example 2 (show internal topology of each node, remove outgroup taxa t9 and t8, and reroot in t1 in both trees)
sharedNodes (tree1, tree2, composition=T, outgroup=c("t9", "t8"), root="t1",)

# Example 3 (plot  two trees)
sharedNodes (tree1, tree2, plotTrees=T, tree.width=8, tree.height=8, tree.fsize=3, tree.adj=c(-1.5,0.5), tree.cex=0.6)

# Example 4 (plot tangle trees)
sharedNodes (tree1, tree2, tanglegram=T, tanglegram.margin=7, tanglegram.lab.cex=0.5)

}
\author{
Daniel YM Nakamura, Taran Grant
}
