% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniqueNodes.R
\name{uniqueNodes}
\alias{uniqueNodes}
\title{uniqueNodes}
\usage{
uniqueNodes(
  tree1,
  tree2,
  composition = T,
  outgroup = NULL,
  root = NULL,
  dataframe = F
)
}
\arguments{
\item{tree1}{A .phylo tree that can be loaded using ape::read.tree for NEWICK files or TreeTools::ReadTntTree for TNT files}

\item{tree2}{Another .phylo tree}

\item{composition}{Optional. Specify if composition of corresponding clades should be present in the dataframe (by default, composition = T)}

\item{outgroup}{Optional. Specify outgroup taxa to remove (by default, outgroup = F assumes that the user does not want to remove outgroup taxa)}

\item{root}{Optional. Specify the same root for both trees, which is recommended to facilitate tree comparisons (by default, root = F assumes that trees share the same root)}

\item{dataframe}{Optional. Write a TSV file in current directory containing the output dataframe (by default, dataset = T).}
}
\description{
Creates two separate dataframes containing the list of descendants and support values of unique clades between two trees.
}
\examples{
# Example 1 (identify unique nodes)
tree1 = read.tree (text="(t1,(t2,(t3,(t4,t5)75)32)45);")
tree2 = read.tree (text="(t1,(t6,(t3,(t4,t5)47)53)94);")
uniqueNodes (tree1, tree2)

# Example 2 (count unique nodes)
tree1 = read.tree (text="(t1,(t2,(t3,(t4,t5)75)32)45);")
tree2 = read.tree (text="(t1,(t6,(t3,(t4,t5)47)53)94);")
nrow (uniqueNodes (tree1, tree2, composition = F))

}
\author{
Daniel YM Nakamura, Taran Grant
}
